+++ tsconfig.json
@@ -31,7 +31,7 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["./*"],
-      "@components/*": ["src/components/components/*"],
+      "@common/*": ["src/components/common/*"],
       "@features/*": ["src/components/features/*"],
       "@pages/*": ["src/components/pages/*"],
       "@definitions/*": ["src/components/definitions/*"],
